% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colour_picker.R
\name{mod_colour_picker_server}
\alias{mod_colour_picker_server}
\title{colour_picker Server Functions}
\usage{
mod_colour_picker_server(id, df, colname, palette)
}
\arguments{
\item{id}{module ID}

\item{df}{dataset (data.frame; reactive)}

\item{colname}{column in \code{df} to extract unique levels from (string; reactive)}

\item{palette}{default colours to use. If length is less than the size of the    (character)}
}
\value{
a named vector where names are all unique values of column \code{colname} in dataset \code{df} and values are colours based on pickerInput selections.
}
\description{
Generate a reactive set of colourPicker UI elements (1 per unique value in \code{df[colname]}) and
return a palette vector with colours for each unique value.
}
